Separate general hashing methods from table implementation
