+++ osquery/carver/carver.cpp
@@ -24,11 +24,11 @@
 
 #include "osquery/carver/carver.h"
 #include "osquery/core/conversions.h"
+#include "osquery/core/hash.h"
 #include "osquery/core/json.h"
 #include "osquery/filesystem/fileops.h"
 #include "osquery/remote/serializers/json.h"
 #include "osquery/remote/utility.h"
-#include "osquery/tables/system/hash.h"
 
 namespace fs = boost::filesystem;
 namespace pt = boost::property_tree;
