+++ osquery/carver/tests/carver_tests.cpp
@@ -19,9 +19,9 @@
 #include <osquery/sql.h>
 
 #include "osquery/carver/carver.h"
+#include "osquery/core/hash.h"
 #include "osquery/core/json.h"
 #include "osquery/filesystem/fileops.h"
-#include "osquery/tables/system/hash.h"
 #include "osquery/tests/test_util.h"
 
 namespace osquery {
