+++ osquery/sql/sqlite_hashing.cpp
@@ -12,7 +12,7 @@
 #include <string>
 
 #include "osquery/core/conversions.h"
-#include "osquery/tables/system/hash.h"
+#include "osquery/core/hash.h"
 
 #include <sqlite3.h>
