+++ osquery/tables/events/event_utils.cpp
@@ -11,8 +11,8 @@
 #include <osquery/events.h>
 #include <osquery/sql.h>
 
+#include "osquery/core/hash.h"
 #include "osquery/tables/events/event_utils.h"
-#include "osquery/tables/system/hash.h"
 
 namespace osquery {
