+++ osquery/tables/sleuthkit/sleuthkit.cpp
@@ -21,7 +21,7 @@
 #include <osquery/tables.h>
 
 #include "osquery/core/conversions.h"
-#include "osquery/tables/system/hash.h"
+#include "osquery/core/hash.h"
 
 namespace fs = boost::filesystem;
