+++ osquery/tables/system/darwin/acpi_tables.cpp
@@ -14,7 +14,7 @@
 #include <osquery/tables.h>
 
 #include "osquery/core/conversions.h"
-#include "osquery/tables/system/hash.h"
+#include "osquery/core/hash.h"
 
 namespace osquery {
 namespace tables {
