+++ osquery/tables/system/hash_table.cpp
@@ -8,17 +8,6 @@
  *  You may select, at your option, one of the above-listed licenses.
  */
 
-#include <algorithm>
-#include <iomanip>
-#include <set>
-#include <sstream>
-#include <thread>
-#include <unordered_map>
-#include <vector>
-
-#include <errno.h>
-#include <string.h>
-
 // clang-format off
 #include <sys/types.h>
 #include <sys/stat.h>
@@ -28,18 +17,18 @@
 #include <unistd.h>
 #endif
 
-#include <openssl/md5.h>
-#include <openssl/sha.h>
+#include <set>
+#include <thread>
 
 #include <boost/filesystem.hpp>
 
-#include <osquery/core.h>
-#include <osquery/filesystem.h>
 #include <osquery/flags.h>
-#include <osquery/logger.h>
+#include <osquery/filesystem.h>
 #include <osquery/tables.h>
+#include <osquery/logger.h>
 
-#include "osquery/tables/system/hash.h"
+
+#include "osquery/core/hash.h"
 
 namespace osquery {
 
@@ -55,124 +44,10 @@ HIDDEN_FLAG(uint32,
             20,
             "Number of milliseconds to delay after hashing");
 
+namespace tables {
+
 /// Clear this amount of rows every time cache eviction is triggered.
-#define HASH_CACHE_EVICT_SIZE 5
-
-/// The buffer read size from file IO to hashing structures.
-#define HASH_CHUNK_SIZE 4096
-
-Hash::~Hash() {
-  if (ctx_ != nullptr) {
-    free(ctx_);
-  }
-}
-
-Hash::Hash(HashType algorithm) : algorithm_(algorithm) {
-  if (algorithm_ == HASH_TYPE_MD5) {
-    length_ = MD5_DIGEST_LENGTH;
-    ctx_ = static_cast<MD5_CTX*>(malloc(sizeof(MD5_CTX)));
-    MD5_Init(static_cast<MD5_CTX*>(ctx_));
-  } else if (algorithm_ == HASH_TYPE_SHA1) {
-    length_ = SHA_DIGEST_LENGTH;
-    ctx_ = static_cast<SHA_CTX*>(malloc(sizeof(SHA_CTX)));
-    SHA1_Init(static_cast<SHA_CTX*>(ctx_));
-  } else if (algorithm_ == HASH_TYPE_SHA256) {
-    length_ = SHA256_DIGEST_LENGTH;
-    ctx_ = static_cast<SHA256_CTX*>(malloc(sizeof(SHA256_CTX)));
-    SHA256_Init(static_cast<SHA256_CTX*>(ctx_));
-  } else {
-    throw std::domain_error("Unknown hash function");
-  }
-}
-
-void Hash::update(const void* buffer, size_t size) {
-  if (algorithm_ == HASH_TYPE_MD5) {
-    MD5_Update(static_cast<MD5_CTX*>(ctx_), buffer, size);
-  } else if (algorithm_ == HASH_TYPE_SHA1) {
-    SHA1_Update(static_cast<SHA_CTX*>(ctx_), buffer, size);
-  } else if (algorithm_ == HASH_TYPE_SHA256) {
-    SHA256_Update(static_cast<SHA256_CTX*>(ctx_), buffer, size);
-  }
-}
-
-std::string Hash::digest() {
-  std::vector<unsigned char> hash;
-  hash.assign(length_, '\0');
-
-  if (algorithm_ == HASH_TYPE_MD5) {
-    MD5_Final(hash.data(), static_cast<MD5_CTX*>(ctx_));
-  } else if (algorithm_ == HASH_TYPE_SHA1) {
-    SHA1_Final(hash.data(), static_cast<SHA_CTX*>(ctx_));
-  } else if (algorithm_ == HASH_TYPE_SHA256) {
-    SHA256_Final(hash.data(), static_cast<SHA256_CTX*>(ctx_));
-  }
-
-  // The hash value is only relevant as a hex digest.
-  std::stringstream digest;
-  for (size_t i = 0; i < length_; i++) {
-    digest << std::hex << std::setw(2) << std::setfill('0') << (int)hash[i];
-  }
-
-  return digest.str();
-}
-
-std::string hashFromBuffer(HashType hash_type,
-                           const void* buffer,
-                           size_t size) {
-  Hash hash(hash_type);
-  hash.update(buffer, size);
-  return hash.digest();
-}
-
-MultiHashes hashMultiFromFile(int mask, const std::string& path) {
-  std::map<HashType, std::shared_ptr<Hash>> hashes = {
-      {HASH_TYPE_MD5, std::make_shared<Hash>(HASH_TYPE_MD5)},
-      {HASH_TYPE_SHA1, std::make_shared<Hash>(HASH_TYPE_SHA1)},
-      {HASH_TYPE_SHA256, std::make_shared<Hash>(HASH_TYPE_SHA256)},
-  };
-
-  auto s = readFile(path,
-                    0,
-                    HASH_CHUNK_SIZE,
-                    false,
-                    true,
-                    ([&hashes, &mask](std::string& buffer, size_t size) {
-                      for (auto& hash : hashes) {
-                        if (mask & hash.first) {
-                          hash.second->update(&buffer[0], size);
-                        }
-                      }
-                    }),
-                    true);
-
-  MultiHashes mh = {};
-  if (!s.ok()) {
-    return mh;
-  }
-
-  mh.mask = mask;
-  if (mask & HASH_TYPE_MD5) {
-    mh.md5 = hashes.at(HASH_TYPE_MD5)->digest();
-  }
-  if (mask & HASH_TYPE_SHA1) {
-    mh.sha1 = hashes.at(HASH_TYPE_SHA1)->digest();
-  }
-  if (mask & HASH_TYPE_SHA256) {
-    mh.sha256 = hashes.at(HASH_TYPE_SHA256)->digest();
-  }
-  return mh;
-}
-
-std::string hashFromFile(HashType hash_type, const std::string& path) {
-  auto hashes = hashMultiFromFile(hash_type, path);
-  if (hash_type == HASH_TYPE_MD5) {
-    return hashes.md5;
-  } else if (hash_type == HASH_TYPE_SHA1) {
-    return hashes.sha1;
-  } else {
-    return hashes.sha256;
-  }
-}
+const size_t kHashCacheEvictSize{5};
 
 /**
  * @brief Implements persistent in-memory caching of files' hashes.
@@ -266,7 +141,7 @@ bool FileHashCache::load(const std::string& path, MultiHashes& out) {
   if (entry == cache.end()) { // none, load
     if (cache.size() >= FLAGS_hash_cache_max) {
       // too large, evict
-      for (size_t i = 0; i < HASH_CACHE_EVICT_SIZE; ++i) {
+      for (size_t i = 0; i < kHashCacheEvictSize; ++i) {
         if (lru.empty()) {
           continue;
         }
@@ -343,8 +218,6 @@ void genHashForFile(const std::string& path,
   results.push_back(r);
 }
 
-namespace tables {
-
 QueryData genHash(QueryContext& context) {
   QueryData results;
   boost::system::error_code ec;
@@ -417,5 +290,5 @@ QueryData genHash(QueryContext& context) {
 
   return results;
 }
-}
-}
+} // namespace tables
+} // namespace osquery
