+++ osquery/tables/system/posix/smbios_utils.cpp
@@ -9,7 +9,7 @@
  */
 
 #include "osquery/tables/system/smbios_utils.h"
-#include "osquery/tables/system/hash.h"
+#include "osquery/core/hash.h"
 
 namespace osquery {
 namespace tables {
